+++ icu4c/source/i18n/unesctrn.cpp
@@ -87,7 +87,7 @@ static UChar* copySpec(const UChar* spec) {
     UChar *result = (UChar *)uprv_malloc(len*sizeof(UChar));
     // Check for memory allocation error. 
     if (result != NULL) {
-    	uprv_memcpy(result, spec, len*sizeof(result[0]));
+    	uprv_memcpy(result, spec, (size_t)len*sizeof(result[0]));
     }
     return result;
 }
